+++ client/scripts/actions/ColorActions.js
@@ -18,8 +18,4 @@
 
 import alt from '../alt';
 
-class _ColorActions {
-  setColorBlindMode(newValue) { this.dispatch(newValue); }
-}
-
-export default alt.createActions(_ColorActions);
+export default alt.generateActions('setColorBlindMode');