+++ client/scripts/actions/ConfigActions.js
@@ -18,190 +18,59 @@
 
 import alt from '../alt';
 
-class _ConfigActions {
-  constructor() {
-    this.generateActions(
-      'toggleAutoApprove',
-      'setStateForTest'
-    );
-  }
-  startEditingDeclarationType(id) { this.dispatch(id); }
-
-  updateDeclarationType(id, newValue) {
-    this.dispatch({id, newValue});
-  }
-
-  stopEditingDeclarationType(id) { this.dispatch(id); }
-
-  toggleDeclarationType(id) { this.dispatch(id); }
-
-  startEnteringNewDeclarationType() { this.dispatch(); }
-
-  deleteInProgressCustomDeclarationType() { this.dispatch(); }
-
-  saveNewDeclarationType() { this.dispatch(); }
-
-  setNewDeclarationTypeText(newValue) { this.dispatch(newValue); }
-
-  deleteDeclarationType(id) { this.dispatch(id); }
-
-  enableDisclosureType(typeCd) { this.dispatch(typeCd); }
-
-  disableDisclosureType(typeCd) { this.dispatch(typeCd); }
-
-  updateDisclosureType(typeCd, newValue) {
-    this.dispatch({typeCd, newValue});
-  }
-
-  enableSponsorLookup() { this.dispatch(); }
-
-  disableSponsorLookup() { this.dispatch(); }
-
-  setDueDate(newDate) { this.dispatch(newDate); }
-
-  setIsRollingDueDate(value) { this.dispatch(value); }
-
-  setWarningValueOnNotification(id, newValue) {
-    this.dispatch({id, newValue});
-  }
-
-  setWarningPeriodOnNotification(id, newValue) {
-    this.dispatch({id, newValue});
-  }
-
-  setReminderTextOnNotification(id, newValue) {
-    this.dispatch({id, newValue});
-  }
-
-  saveNewNotification() { this.dispatch(); }
-
-  deleteNotification(id) { this.dispatch(id); }
-
-  questionTypeChosen(category, questionId, type) {
-    this.dispatch({
-      questionId,
-      type,
-      category
-    });
-  }
-
-  questionTextChanged(category, questionId, text) {
-    this.dispatch({questionId, text, category});
-  }
-
-  updateQuestions(category, questions) {
-    this.dispatch({questions, category});
-  }
-
-  cancelNewQuestion(category) { this.dispatch({category}); }
-
-  saveNewQuestion(category) { this.dispatch({category}); }
-
-  startNewQuestion(category) { this.dispatch({category}); }
-
-  deleteQuestion(category, questionId) {
-    this.dispatch({questionId, category});
-  }
-
-  saveQuestionEdit(category, questionId) {
-    this.dispatch({questionId, category});
-  }
-
-  startEditingQuestion(category, questionId) {
-    this.dispatch({questionId, category});
-  }
-
-  cancelQuestionEdit(category, questionId) {
-    this.dispatch({questionId, category});
-  }
-
-  criteriaChanged(category, questionId, newValue) {
-    this.dispatch({
-      questionId,
-      newValue,
-      category
-    });
-  }
-
-  multiSelectOptionAdded(category, questionId, newValue) {
-    this.dispatch({
-      questionId,
-      newValue,
-      category
-    });
-  }
-
-  multiSelectOptionDeleted(category, questionId, optionId) {
-    this.dispatch({
-      questionId,
-      optionId,
-      category
-    });
-  }
-
-  requiredNumSelectionsChanged(category, questionId, newValue) {
-    this.dispatch({
-      questionId,
-      newValue,
-      category
-    });
-  }
-
-  relationshipPeopleChanged(newPeople) {
-    this.dispatch(newPeople);
-  }
-
-  relationshipPeopleEnabled(newValue) {
-    this.dispatch(newValue);
-  }
-
-  enabledChanged(typeCd, newValue) {
-    this.dispatch({typeCd, newValue});
-  }
-
-  typeEnabledChanged(typeCd, newValue) {
-    this.dispatch({typeCd, newValue});
-  }
-
-  amountEnabledChanged(typeCd, newValue) {
-    this.dispatch({typeCd, newValue});
-  }
-
-  typeOptionsChanged(typeCd, newList) {
-    this.dispatch({typeCd, newList});
-  }
-
-  amountOptionsChanged(typeCd, newList) {
-    this.dispatch({typeCd, newList});
-  }
-
-  destinationEnabledChanged(typeCd, newValue) {
-    this.dispatch({typeCd, newValue});
-  }
-
-  dateEnabledChanged(typeCd, newValue) {
-    this.dispatch({typeCd, newValue});
-  }
-
-  reasonEnabledChanged(typeCd, newValue) {
-    this.dispatch({typeCd, newValue});
-  }
-
-  setCertificationText(newText) { this.dispatch(newText); }
-
-  setCertificationRequired(newValue) { this.dispatch(newValue); }
-
-  setInstructions(step, newValue) {
-    this.dispatch({step, newValue});
-  }
-
-  toggleInstructionsExpanded() { this.dispatch(); }
-
-  loadConfig(id) { this.dispatch(id); }
-
-  saveAll() { this.dispatch(); }
-
-  undoAll() { this.dispatch(); }
-}
-
-export default alt.createActions(_ConfigActions);
+export default alt.generateActions(
+  'toggleAutoApprove',
+  'setStateForTest',
+  'startEditingDeclarationType',
+  'stopEditingDeclarationType',
+  'toggleDeclarationType',
+  'startEnteringNewDeclarationType',
+  'deleteInProgressCustomDeclarationType',
+  'saveNewDeclarationType',
+  'setNewDeclarationTypeText',
+  'deleteDeclarationType',
+  'enableDisclosureType',
+  'disableDisclosureType',
+  'enableSponsorLookup',
+  'disableSponsorLookup',
+  'setDueDate',
+  'setIsRollingDueDate',
+  'saveNewNotification',
+  'deleteNotification',
+  'relationshipPeopleChanged',
+  'relationshipPeopleEnabled',
+  'setCertificationText',
+  'setCertificationRequired',
+  'toggleInstructionsExpanded',
+  'loadConfig',
+  'saveAll',
+  'undoAll',
+  'cancelNewQuestion',
+  'saveNewQuestion',
+  'startNewQuestion',
+  'updateDeclarationType',
+  'updateDisclosureType',
+  'setWarningValueOnNotification',
+  'setWarningPeriodOnNotification',
+  'setReminderTextOnNotification',
+  'questionTypeChosen',
+  'questionTextChanged',
+  'updateQuestions',
+  'deleteQuestion',
+  'saveQuestionEdit',
+  'startEditingQuestion',
+  'cancelQuestionEdit',
+  'criteriaChanged',
+  'multiSelectOptionAdded',
+  'multiSelectOptionDeleted',
+  'requiredNumSelectionsChanged',
+  'enabledChanged',
+  'typeEnabledChanged',
+  'amountEnabledChanged',
+  'typeOptionsChanged',
+  'amountOptionsChanged',
+  'destinationEnabledChanged',
+  'dateEnabledChanged',
+  'reasonEnabledChanged',
+  'setInstructions'
+);