+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+##coi-1602
+* Fixed bug on the pi-revise screen which prevented adding a new relation to an existing entity while reviewing.
+
 ##coi-1601
 * Now tracks the user id, user name, and time of change when a config modification is made.  DB MIGRATION REQUIRED!
 * Instruction box now spans full width of the pane.